telemetry documentation + minor fixes

+++ NativeCode~/projects/game_telemetry/src/events/DeathEvent.cs
@@ -18,11 +18,11 @@ namespace game_telemetry
         public enum DeathType { FELL, DAMAGE };
 
         [JsonProperty(Order = 4)] public DeathType Death { get; private set; }
-        [JsonProperty(Order = 5)] public int PosPlayerX { get; private set; }
-        [JsonProperty(Order = 6)] public int PosPlayerY { get; private set; }
+        [JsonProperty(Order = 5)] public float PosPlayerX { get; private set; }
+        [JsonProperty(Order = 6)] public float PosPlayerY { get; private set; }
         [JsonProperty(Order = 7)] public int ZoneID { get; private set; }
 
-        public DeathEvent(EventType type, int posPlayerX, int posPlayerY, DeathType death, int zoneID) : base(type)
+        public DeathEvent(EventType type, float posPlayerX, float posPlayerY, DeathType death, int zoneID) : base(type)
         {
             Death = death;
             PosPlayerX = posPlayerX;
